--- a/openvdb_houdini/openvdb_houdini/SOP_OpenVDB_Visualize.cc
+++ b/openvdb_houdini/openvdb_houdini/SOP_OpenVDB_Visualize.cc
@@ -1843,7 +1843,16 @@ TreeVisualizer::allocateOffsetArrays(const GridType& grid)
 {
     // allocate offsets per node arrays - no zero value initialization
 
-    auto nodeCounts = grid.tree().nodeCount();
+#if OPENVDB_ABI_VERSION_NUMBER >= 7
+    const auto nodeCounts = grid.tree().nodeCount();
+#else
+    // Tree::nodeCount() was only added in ABI=7, so use a NodeIterator with an earlier ABI
+    std::vector<openvdb::Index32> nodeCounts(grid.tree().treeDepth());
+    for (auto it = grid.tree().cbeginNode(); it; ++it) {
+        ++(nodeCounts[it.getLevel()]);
+    }
+#endif
+
     for (const auto& count : nodeCounts) {
         mPointOffsets.emplace_back(new size_t[count]);
         mVertexOffsets.emplace_back(new size_t[count]);