--- a/pendingchanges/vdb_visualize_sop.txt
+++ b/pendingchanges/vdb_visualize_sop.txt
@@ -1,0 +1,3 @@
+Houdini:
+    - VDB Visualize Tree SOP is now multi-threaded and provides slicing and
+      color remapping.