A fairly significant re-write of the VDB Visualize Tree SOP. Here are the changes summarized:

* Remove logic for DWA-specific surfacing and region of interest (parameters are now obsolete)
* Use multi-threading for the geometry generation
* Introduce slicing and color remapping
* Add gl_lit=0 detail attribute to disable gl lighting

**Note:** care has been taken to preserve the geometry and point ordering such that the output is identical to what it was before. For instance, there is still two-pass geometry generation where the leaf nodes and internal nodes are processed first and then the voxels and tiles (even though a simpler implementation would likely have been possible by combining them).

In terms of multi-threading performance, here's the full res Disney cloud asset visualized with leaf nodes, internal nodes and tiles as wireframe box and points as voxels:

Old implementation - 1 minute 9 seconds
New implementation - 13.8 seconds

The main change here is moving away from the "addPoint()" and "addBox()" calls to a method where the offsets are pre-computed and the point, vertex and polygon data generated in parallel. The only caveat is that the polygon generation is still single-threaded for now where the solid boxes and wire-frame boxes are mixed as part of the same pass (ie internal nodes shown as wireframe, leaf nodes as solid). 

The slicing has been pretty closely modeled on the Volume Slice SOP where I tried to keep the same parameter names and layout (see parameter pane below). There are however three modes for slicing - one where only geometry that intersects a plane is generated, one where all voxels in an intersecting leaf are generated (sometimes useful visually) and finally one where the geometry is "flattened" to the plane so that a 3D box becomes a 2D square more akin to the output of the Volume Slice SOP.

Here is a side-by-side of a slice through the full res Disney cloud asset:

![volumeslice](https://user-images.githubusercontent.com/5251654/127361129-4da7dab1-140e-4c6b-b3c8-20f6a0d1e702.png)

The left is the Volume Slice SOP with regular point spacing, the right is the VDB Visualize Tree SOP with geometry that matches that precise structure of the tree. Performance for this example:

Volume Slice SOP - 1.64s
VDB Visualize Tree SOP - 0.09s

The huge performance increase is primarily from avoiding the need to interpolate the VDB. Even with some very large VDBs, it's now possible to change the offset slider and have it update in real-time.

The offset defaults to a relative -1..1 range, however there is a "use world-space" toggle which changes the value to Houdini units relative to the VDB origin. The offsets are stored in the Node cache so providing the geometry has been generated once, clicking the toggle will translate the value seamlessly between the two - the key advantage being that the world space mode maps to the same plane even when the bounds of the VDB is changing (the node cache approach is similar to one used already in VDB from Particles SOP).

Finally, here's the new parameter pane:

![vdbvisualize](https://user-images.githubusercontent.com/5251654/127362461-5e09e1ee-2a77-4535-b72a-1a5cd26f48f8.png)

The new parameters are all added to the bottom, but I'm open to changing the layout to make it more usable.