@jmlait 2021-08-17
/home/jlait/dev_green/src/houdini/custom/DWA/vdb/SOP_OpenVDB_Visualize.C:1954:43: warning: suggest parentheses around ‘&&’ within ‘||’ [-Wparentheses]
                  style2 == STYLE_WIRE_BOX && (style1 != STYLE_SOLID_BOX)) {
                  ~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/jlait/dev_green/src/houdini/custom/DWA/vdb/SOP_OpenVDB_Visualize.C:1965:51: warning: suggest parentheses around ‘&&’ within ‘||’ [-Wparentheses]
                         style2 == STYLE_SOLID_BOX && (style1 != STYLE_WIRE_BOX)) {


This seems suspicious, I think the warning is right and style2 should have had a preceding parenthesis?  At least from the structure it looks that way, I've not looked at the actual logic.