@jmlait 2021-08-17
Note on GEO_Detail there are:
   bool        setDetailAttributeI(const UT_StringHolder &aname, int v);
These make very quick ways to just splat stuff like gl_lit without so much code.