@jmlait 2021-08-17
Why is there
          "none", "No Mapping",
when that doesn't link to a valid UT_Color setting?  The switch to test this doesn't include "none".