@jmlait 2021-08-17
Compiled and passes our reg tests after I suppressed gl_lit.  (I don't mind gl_lit being added, btw, just suppressed to make sure the rest matched.)