@danrbailey 2021-08-17
Thanks for those suggestions, I've addressed them now.

> This seems suspicious, I think the warning is right and style2 should have had a preceding parenthesis? At least from the structure it looks that way, I've not looked at the actual logic.

Yeah. I think the code is correct because && precedence is before || but I agree that it's not clear, so I've tidied this up.

> Why is there
> "none", "No Mapping",
> when that doesn't link to a valid UT_Color setting? The switch to test this doesn't include "none".

I don't have the source code of the Volume Slice SOP to look at here, but I was trying to match the implementation as closely as possible. There is not a "No Mapping" option under `UT_Color::getRampColor()` even though there is in the drop-down of the Volume Slice SOP, so instead the default value it uses (when "No Mapping" is selected) is UT_COLORRAMP_NUMMODES which appears to produce the correct output?