@danrbailey 2021-08-17
> NONE is a special case to directly copy the value into Cd, which also implies no clamping. NUMMODES triggers a hundred asserts for me as the lower level libraries are not happy, but try to fail-safe to something reasonable :>
> 
> ```
>                 // Find the colour correlating to this value.
>                 if (sopparms.getVismode() == SOP_VolumeSliceEnums::Vismode::NONE)
>                 {
>                     cd = v;
>                 }
>                 else if (color_ramp)
>                 {
>                     float vals[4];
>                     color_ramp->rampLookup(v, vals);
>                     cd.assign(vals[0], vals[1], vals[2]);
>                 }
>                 else
>                 {
>                     UT_Color::getRampColor((UT_ColorRamp) (int(sopparms.getVismode())-1), v,
>                                     &cd.x(), &cd.y(), &cd.z());
>                 }
> ```
> 
> I'd actually suggest going straight to a Ramp and not passing UT_ColorRamp at all. We've since added all the standard vis as presets in the color ramp so it feels less necessary to have that specific drop down. Volume Slice only has it because it predated having defaults on colour ramps. But if a full ramp seems overkill, I won't press the point either.

Full ramp sounds fine to me, I'll make that change. Thanks for explaining.