@danrbailey 2021-08-18
@jmlait - I've made the changes to switch to a Ramp as you suggested, but I'm not sure that I've done it correctly, I couldn't find many examples of working with ramps in a verbified context. I also couldn't figure out how to apply the "infra-red" preset by default so that it matches the Volume Slice SOP, so I instead opted to just set the "rampcolordefault" spare data with infra-red values. Would you mind reviewing this change please?

https://github.com/AcademySoftwareFoundation/openvdb/pull/1138/commits/c098944d40de14c8dcd52e504b82490261e121d0