@jmlait 2021-08-31
Regarding defaulting to infrared: to my knowledge that is the best way ATM.